Move exists out of companion object
